+++ superduper/__main__.py
@@ -2,15 +2,11 @@ import sys
 
 import click
 
-from superduper.cli import app, apply 
+from superduper.cli import app, main
 
 __all__ = (
-    'apply',
-    'info',
-    'local_cluster',
-    'vector_search',
-    'cdc',
-    'ray_serve',
+    'app',
+    'main',
 )