+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the bug in the CI where plugins are skipping tests.
 - Updated CONTRIBUTING.md
 - Add README.md files for the plugins.
+- Add templates to project
 
 #### New Features & Functionality